+++ docker-compose.yaml
@@ -1,20 +1,32 @@
 version: "3.9"
 
 services:
+  nats:
+    image: nats:latest
+    container_name: nats_server
+    ports:
+      - "4222:4222"
+      - "8222:8222"
+    networks:
+      - app_network
+    restart: unless-stopped
+
   parser:
     build:
       context: ./src/services/parser
       dockerfile: Dockerfile
     container_name: log_parser
     environment:
-      - LOG_DIR=/var/log/logger
-      - PARSED_LOG_DIR=/var/log/parsed
+      - NATS_URL=nats://nats_server:4222
+      - OUTPUT_DIR=/var/log/parsed
     volumes:
-      - ./logs:/var/log/logger           # shared raw logs
       - ./parsed_logs:/var/log/parsed    # shared parsed logs
     networks:
       - app_network
     restart: unless-stopped
+    depends_on:
+      - nats
+
   centrifugo:
     image: centrifugo/centrifugo:v6
     container_name: centrifugo_server
@@ -62,8 +74,11 @@ services:
     depends_on:
       - centrifugo
       - log-generator
+      - nats
     networks:
       - app_network
+    environment:
+      - NATS_URL=nats://nats_server:4222
     ports:
       - "8081:8080"
     restart: unless-stopped
