--- src/services/collector/internal/models/nats.go
@@ -1,5 +1,0 @@
-package models
-
-type NatsConfig struct {
-	URL string `yaml:"url"`
-}
