+++ src/services/collector/main.go
@@ -1,31 +1,24 @@
 package main
 
 import (
+	"collector/services"
 	"log"
 	"os"
 )
-
-type App struct {
-	Config
-}
-
-func NewApp(cfg Config) *App {
-	return &App{Config: cfg}
-}
 
 func main() {
 	logger := log.New(os.Stdout, "[LogCollector] ", log.LstdFlags)
 
 	//Load configuration
-	config, err := loadConfig("/app/config.yml", logger)
+	config, err := services.LoadConfig("/app/config.yml", logger)
 	if err != nil {
 		log.Fatalf("failed to load config: %v", err)
 		return
 	}
 
-	app := NewApp(config)
+	app := services.NewApp(config)
 
-	service, err := NewLogCollectorService(config, logger)
+	service, err := services.NewLogCollectorService(config, logger)
 	if err != nil {
 		log.Fatalf("failed to create collector service: %v", err)
 	}
@@ -36,7 +29,7 @@ func main() {
 		}
 	}()
 
-	server := app.setupRouter()
+	server := app.SetupRouter()
 	logger.Println("Starting Gin server on :8080")
 	if err := server.Run(":8080"); err != nil {
 		log.Fatalf("failed to start Gin server: %v", err)
