+++ src/services/collector/services/app.go
@@ -1,0 +1,9 @@
+package services
+
+type App struct {
+	Config Config
+}
+
+func NewApp(cfg Config) *App {
+	return &App{Config: cfg}
+}
