+++ src/services/collector/services/response.go
@@ -1,4 +1,4 @@
-package main
+package services
 
 import (
 	"net/http"
