+++ src/services/generator/main.go
@@ -4,15 +4,6 @@ import (
 	"log"
 )
 
-type LogEntry struct {
-	Timestamp string `json:"timestamp"`
-	LogType   string `json:"log_type"`
-	UserID    string `json:"user_id"`
-	Duration  int    `json:"duration"`
-	Message   string `json:"message"`
-	RequestID string `json:"request_id"`
-	Service   string `json:"service"`
-}
 
 type App struct {
 	Config
